Minor fixes to alarm clock

+++ threads/thread.h
@@ -93,7 +93,7 @@ struct thread
 
     /* Shared between thread.c and synch.c. */
     struct list_elem elem;              /* List element. */
-    int wake_up_time;                   /* Past this time, wake up. */
+    int64_t wake_up_time;               /* Past this time, wake up. */
 
 #ifdef USERPROG
     /* Owned by userprog/process.c. */
